Basic Rewrite: moved out of ES6 Classes, into functional exports
