+++ tsconfig.json
@@ -3,9 +3,9 @@
     "outDir": "./dist/",
     "noImplicitAny": true,
     "declaration": true,
-    "lib": ["es2017", "es7", "es6", "dom"],
+    "lib": ["es2017", "es7", "dom"],
     "module": "CommonJS",
-    "target": "es5",
+    "target": "es2017",
     "allowJs": true,
     "importsNotUsedAsValues": "preserve",
     "esModuleInterop": true
